third_party/WebKit/LayoutTests/media/video-autoplay.html

<!DOCTYPE html>
<title>Verify that "autoplay" attribute on a media element starts playback automatically.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<video autoplay></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");
    assert_true(video.autoplay);
    assert_not_equals(video.getAttribute("autoplay"), null);

    video.onplay = t.step_func_done(function () {
        assert_false(video.paused);
    });

    video.src = "content/test.ogv";
});
</script>
